+++ src/chat/__tests__/cards/ChatDetails.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import ChatDetails from '../components/cards/ChatDetails';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import ChatDetails from '../../components/cards/ChatDetails';
 
 describe('<ChatDetails />', () => {
   it('does render correctly', () => {
