+++ src/chat/__tests__/chat/Chat.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { waitFor } from '@testing-library/react';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import { getMessages as mockGetMessages } from '../../utils/api';
-import Chat from '../components/chat/Chat';
-import { currentUserMock } from '../testHelpers';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import { getMessages as mockGetMessages } from '../../../utils/api';
+import Chat from '../../components/chat/Chat';
+import { currentUserMock } from '../../testHelpers';
 
-jest.mock('../../utils/api');
+jest.mock('../../../utils/api');
 const onSendMessageMock = jest.fn();
 
 describe('<Chat />', () => {
