+++ src/chat/__tests__/chat/Message.test.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../components/withProviders';
-import { MESSAGE_CONTAINER } from '../../utils/test-helpers';
-import { renderWithRedux } from '../../utils/test-utils';
-import Message from '../components/chat/Message';
-import { testMessage } from '../testHelpers';
+import withProviders from '../../../components/withProviders';
+import { MESSAGE_CONTAINER } from '../../../utils/test-helpers';
+import { renderWithRedux } from '../../../utils/test-utils';
+import Message from '../../components/chat/Message';
+import { testMessage } from '../../testHelpers';
 
 describe('<Message />', () => {
   it('renders received message with proper styles', () => {
