+++ src/chat/__tests__/chatHeader/ChatHeaderLabel.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import ChatHeaderLabel from '../components/chatHeader/ChatHeaderLabel';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import ChatHeaderLabel from '../../components/chatHeader/ChatHeaderLabel';
 
 const label = 'test-label';
