+++ src/chat/__tests__/chatRooms/ChatRoomItem.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import ChatRoomItem from '../components/chatRooms/ChatRoomItem';
-import { chatRoomsMock } from '../testHelpers';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import ChatRoomItem from '../../components/chatRooms/ChatRoomItem';
+import { chatRoomsMock } from '../../testHelpers';
 
 const routePath = `/nutri/inbox`;
 const chatRoomMock = chatRoomsMock[0];
