+++ src/chat/__tests__/chatRooms/ChatRooms.test.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { waitFor } from '@testing-library/react';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import { getChatRooms as mockGetChatRooms } from '../../utils/api';
-import ChatRooms from '../components/chatRooms/ChatRooms';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import { getChatRooms as mockGetChatRooms } from '../../../utils/api';
+import ChatRooms from '../../components/chatRooms/ChatRooms';
 
-jest.mock('../../utils/api');
+jest.mock('../../../utils/api');
 
 describe('<ChatRooms />', () => {
   it('fetches data correctly', async () => {
