+++ src/chat/__tests__/chatRooms/ChatRoomsList.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import ChatRoomsList from '../components/chatRooms/ChatRoomsList';
-import { chatRoomsMock } from '../testHelpers';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import ChatRoomsList from '../../components/chatRooms/ChatRoomsList';
+import { chatRoomsMock } from '../../testHelpers';
 
 describe('<ChatRoomsList />', () => {
   it('renders message and patient name correctly', async () => {
