+++ src/chat/__tests__/inboxSidebar/InboxSidebar.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
 import { waitFor } from '@testing-library/react';
-import withProviders from '../../components/withProviders';
-import { renderWithRedux } from '../../utils/test-utils';
-import InboxSidebar from '../components/inboxSidebar/InboxSidebar';
-import { getInboxesList } from '../../utils/api';
-import { chatConversationsMock } from '../testHelpers';
+import withProviders from '../../../components/withProviders';
+import { renderWithRedux } from '../../../utils/test-utils';
+import InboxSidebar from '../../components/inboxSidebar/InboxSidebar';
+import { getInboxesList } from '../../../utils/api';
+import { chatConversationsMock } from '../../testHelpers';
 
-jest.mock('../../utils/api');
+jest.mock('../../../utils/api');
 const mockGetInboxesList = getInboxesList as jest.Mock;
 
 describe('<InboxSidebar />', () => {
